+++ apps/hodejegeren/build.gradle
@@ -61,10 +61,9 @@ dependencies {
     implementation 'javax.xml.bind:jaxb-api'
     implementation 'org.aspectj:aspectjweaver:1.9.7'
     implementation 'io.micrometer:micrometer-registry-prometheus'
-    implementation 'net.logstash.logback:logstash-logback-encoder:7.2'
 
-    implementation 'io.springfox:springfox-swagger2:3.0.0'
-    implementation 'io.springfox:springfox-swagger-ui:3.0.0'
+    implementation 'org.springdoc:springdoc-openapi-ui:1.6.9'
+    implementation 'net.logstash.logback:logstash-logback-encoder:7.2'
 
     implementation 'no.nav.registre:testnorge-consumers:1.12.0-RELEASE'
     implementation 'com.google.guava:guava:31.0.1-jre'
