+++ apps/hodejegeren/src/main/java/no/nav/testnav/apps/hodejegeren/config/OpenApiConfig.java
@@ -1,0 +1,41 @@
+package no.nav.testnav.apps.hodejegeren.config;
+
+import io.swagger.v3.oas.models.OpenAPI;
+import io.swagger.v3.oas.models.info.Contact;
+import io.swagger.v3.oas.models.info.Info;
+import io.swagger.v3.oas.models.info.License;
+import no.nav.testnav.libs.servletcore.config.ApplicationProperties;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Import;
+import org.springframework.web.servlet.config.annotation.ViewControllerRegistry;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+
+@Configuration
+@Import(ApplicationProperties.class)
+public class OpenApiConfig implements WebMvcConfigurer {
+
+    @Bean
+    public OpenAPI openApi(ApplicationProperties applicationProperties) {
+        return new OpenAPI()
+                .info(new Info()
+                        .title("Testnorge-Hodejegeren")
+                        .version(applicationProperties.getVersion())
+                        .description("Testnorge-Hodejegeren jakter på hoder/testpersoner i TPSF og TPS i testmiljø for å fylle meldingene med identer som oppfyller visse kriterier.")
+                        .termsOfService("https://nav.no")
+                        .contact(new Contact()
+                                .url("https://nav-it.slack.com/archives/CA3P9NGA2")
+                                .email("dolly@nav.no")
+                                .name("Team Dolly")
+                        )
+                        .license(new License()
+                                .name("MIT License")
+                                .url("https://opensource.org/licenses/MIT")
+                        ));
+    }
+
+    @Override
+    public void addViewControllers(ViewControllerRegistry registry) {
+        registry.addViewController("/swagger").setViewName("redirect:/swagger-ui.html");
+    }
+}
