--- apps/hodejegeren/src/main/java/no/nav/testnav/apps/hodejegeren/config/SwaggerConfig.java
@@ -1,64 +1,0 @@
-package no.nav.testnav.apps.hodejegeren.config;
-
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.servlet.config.annotation.ViewControllerRegistry;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-import springfox.documentation.annotations.ApiIgnore;
-import springfox.documentation.builders.ApiInfoBuilder;
-import springfox.documentation.builders.PathSelectors;
-import springfox.documentation.builders.RequestHandlerSelectors;
-import springfox.documentation.service.ApiInfo;
-import springfox.documentation.service.Contact;
-import springfox.documentation.spi.DocumentationType;
-import springfox.documentation.spring.web.plugins.Docket;
-import springfox.documentation.swagger2.annotations.EnableSwagger2;
-
-import java.util.Collections;
-import java.util.HashSet;
-import java.util.Set;
-
-/**
- * Configure automated swagger API documentation
- */
-
-@Configuration
-@EnableSwagger2
-public class SwaggerConfig implements WebMvcConfigurer {
-
-    @Value("${application.version}")
-    private String appVersion;
-
-    @Bean
-    public Docket api() {
-        Set<String> contentTypeJson = new HashSet<>(Collections.singletonList("application/json"));
-        return new Docket(DocumentationType.SWAGGER_2)
-                .ignoredParameterTypes(ApiIgnore.class)
-                .select()
-                .apis(RequestHandlerSelectors.any())
-                .paths(PathSelectors.ant("/api/**"))
-                .build()
-                .apiInfo(apiInfo())
-                .produces(contentTypeJson)
-                .consumes(contentTypeJson)
-                .useDefaultResponseMessages(false);
-    }
-
-    private ApiInfo apiInfo() {
-        return new ApiInfoBuilder()
-                .title("Testnorge-Hodejegeren")
-                .description("Testnorge-Hodejegeren jakter på hoder/testpersoner i TPSF og TPS i testmiljø for å fylle meldingene med identer som oppfyller visse kriterier.")
-                .version(appVersion)
-                .termsOfServiceUrl("https://nav.no")
-                .contact(new Contact("Fellesregistrene på NAV", "http://stash.devillo.no/projects/FEL/repos/testnorge-hodejegeren/browse", null))
-                .license("Super Strict Licence")
-                .licenseUrl("https://opensource.org/licenses/super-strict-license")
-                .build();
-    }
-
-    @Override
-    public void addViewControllers(ViewControllerRegistry registry) {
-        registry.addViewController("/swagger").setViewName("redirect:/swagger-ui.html");
-    }
-}
